third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-4/do-not-rebuild-styles-on-every-change-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests show that ssp isn't rebuild on every dom mutation


Running: setupTest

Running: testModifySibling
Requested StyleSidebarPane update

Running: testModifySiblingChild

Running: testModifyParent
Requested StyleSidebarPane update

Running: testModifyChild
